+++ prettier.config.mjs
@@ -12,6 +12,7 @@ const config = {
     plugins: ["@1adybug/prettier"],
     controlStatementBraces: "add",
     multiLineBraces: "add",
+    nodeProtocol: "add",
 }
 
 export default config